requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
PyYAML==6.0.2
aiohttp==3.10.10
aiosqlite==0.20.0
feedparser==6.0.11
PyNaCl==1.5.0
yt-dlp==2025.01.15
ddgs==9.6.0
